Disconnect WebSocket once exception is raised
For example, on Android a `SocketException` will be raised when there's no available network. Without this change, the `receiverJob` would continuously try to read messages from the unusable socket.

Rather than handling a `SocketException` only, this change ensures that the WebSocket is not read from anymore once any exception is encountered. The likelihood of a later read attempt working seems to be slim/impossible.